[Python] Fix __hash__ to return native int for Python 3.14 compatibility
combine_hash_codes and array_hash returned int32 (Fable's Int32, not a
subclass of native int), and TypeInfo.__hash__ passed that through.
Python 3.14 strictly requires __hash__ to return a native int. Change
combine_hash_codes to operate on native int throughout, matching the
pattern already established in HashableBase.__hash__.

+++ src/fable-library-py/fable_library/reflection.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass
 from typing import Any, cast
 
 from .array_ import Array
-from .core import FSharpRef, int32
+from .core import FSharpRef
 from .record import Record
 from .types import IntegerTypes
 from .union import Union
@@ -55,8 +55,8 @@ class TypeInfo:
         return equals(self, other)
 
     def __hash__(self) -> int:
-        hashes: list[int32] = [int32(hash(x)) for x in self.generics or []]
-        hashes.append(int32(hash(self.fullname)))
+        hashes: list[int] = [hash(x) for x in self.generics or []]
+        hashes.append(hash(self.fullname))
         return combine_hash_codes(hashes)
